requirements.txt
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.1.0
black==19.3b0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
Django==2.1
django-cors-middleware==1.3.1
django-extensions==2.1.6
django-pytest==0.2.0
djangorestframework==3.9.2
drf-yasg==1.15.0
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
pluggy==0.9.0
psycopg2-binary==2.8.2
py==1.8.0
PyJWT==1.7.1
pytest==4.4.1
pytest-cov==2.6.1
pytest-django==3.4.8
pytz==2019.1
PyYAML==5.1
requests==2.21.0
ruamel.yaml==0.15.94
six==1.10.0
toml==0.10.0
uritemplate==3.0.0
urllib3==1.24.2
whitenoise==4.1.2
